Creando TicketReposistory,TecnicoRepository y TicektDao,TicketScreen,TickeListScreen, TecnicoScreen,TecnicoListScreen
