+++ app/src/main/java/edu/ucne/registrotecnicos/data/local/database/TecnicoDb.kt
@@ -3,15 +3,19 @@ package edu.ucne.registrotecnicos.data.local.database
 import androidx.room.Database
 import androidx.room.RoomDatabase
 import edu.ucne.registrotecnicos.data.local.dao.TecnicoDao
+import edu.ucne.registrotecnicos.data.local.dao.TicketDao
 import edu.ucne.registrotecnicos.data.local.entities.TecnicoEntity
+import edu.ucne.registrotecnicos.data.local.entities.TicketEntity
 
 @Database(
     entities = [
-        TecnicoEntity::class
-    ],
-    version = 1,
+        TecnicoEntity::class,
+        TicketEntity::class],
+
+    version = 3,
     exportSchema = false
 )
 abstract class TecnicoDb : RoomDatabase(){
     abstract fun tecnicoDao(): TecnicoDao
+    abstract fun ticketDao(): TicketDao
 }
