+++ app/src/main/java/edu/ucne/registrotecnicos/data/local/repositories/TecnicoRepository.kt
@@ -1,4 +1,21 @@
 package edu.ucne.registrotecnicos.data.local.repositories
 
-class TecnicoRepository {
+import edu.ucne.registrotecnicos.data.local.database.TecnicoDb
+import edu.ucne.registrotecnicos.data.local.entities.TecnicoEntity
+import kotlinx.coroutines.flow.Flow
+
+class TecnicoRepository(
+    private val tecnicoDb: TecnicoDb
+) {
+    suspend fun saveTecnico(tecnico: TecnicoEntity){
+        tecnicoDb.tecnicoDao().save(tecnico)
+    }
+
+    suspend fun find(id: Int): TecnicoEntity?{
+        return tecnicoDb.tecnicoDao().find(id)
+    }
+
+    fun getAll(): Flow<List<TecnicoEntity>> {
+        return tecnicoDb.tecnicoDao().getAll()
+    }
 }
