--- app/src/main/java/edu/ucne/registrotecnicos/data/local/repositories/TicketReposistory.kt
@@ -1,4 +1,0 @@
-package edu.ucne.registrotecnicos.data.local.repositories
-
-class TicketReposistory {
-}
