+++ app/src/main/java/edu/ucne/registrotecnicos/data/local/repositories/TicketRepository.kt
@@ -1,0 +1,21 @@
+package edu.ucne.registrotecnicos.data.local.repositories
+
+import edu.ucne.registrotecnicos.data.local.database.TecnicoDb
+import edu.ucne.registrotecnicos.data.local.entities.TicketEntity
+import kotlinx.coroutines.flow.Flow
+
+class TicketRepository(
+    private val tecnicoDb: TecnicoDb
+){
+    suspend fun saveTicket(ticket: TicketEntity){
+        tecnicoDb.ticketDao().save(ticket)
+    }
+
+    suspend fun find(id: Int): TicketEntity?{
+        return tecnicoDb.ticketDao().find(id)
+    }
+
+    suspend fun getAll(): Flow<List<TicketEntity>> {
+        return tecnicoDb.ticketDao().getAll()
+    }
+}
