+++ app/src/main/java/edu/ucne/registrotecnicos/presentation/navigation/Screen.kt
@@ -2,9 +2,10 @@ package edu.ucne.registrotecnicos.presentation.navigation
 
 import kotlinx.serialization.Serializable
 
+
 sealed class Screen {
     @Serializable
-    data object  TecnicoList : Screen()
+    data object TecnicoList : Screen()
     @Serializable
     data class Tecnico(val tecnicoId: Int) : Screen()
 }
